requirements.txt
scikit-learn
ipykernel
pandas
matplotlib
imbalanced-learn
pyaml
mlflow